ch(structure) - Move models to separate file.
